New INT32 and INT64 dense blocks, incl refactoring and tests

+++ src/main/java/org/tugraz/sysds/hops/DataOp.java
@@ -378,7 +378,7 @@ public class DataOp extends Hop
 		{
 			switch( getValueType() ) 
 			{
-				case INT:
+				case INT64:
 					ret = OptimizerUtils.INT_SIZE; break;
 				case FP64:
 					ret = OptimizerUtils.DOUBLE_SIZE; break;
